requirements.txt
# Flask==2.3.2
# flask-cors==3.0.10
# shazamio==0.4.0.1
# python-dotenv==1.0.0
flask
flask_cors
shazamio
asyncio
python-dotenv

